+++ docs/html/search/searchdata.js
@@ -4,7 +4,8 @@ var indexSectionsWithContent =
   1: "c",
   2: "c",
   3: "c",
-  4: "a"
+  4: "c",
+  5: "a"
 };
 
 var indexSectionNames =
@@ -12,8 +13,9 @@ var indexSectionNames =
   0: "all",
   1: "files",
   2: "functions",
-  3: "typedefs",
-  4: "pages"
+  3: "variables",
+  4: "typedefs",
+  5: "pages"
 };
 
 var indexSectionLabels =
@@ -21,7 +23,8 @@ var indexSectionLabels =
   0: "All",
   1: "Files",
   2: "Functions",
-  3: "Typedefs",
-  4: "Pages"
+  3: "Variables",
+  4: "Typedefs",
+  5: "Pages"
 };