+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
 * Unreleased
+* 0.2 (2020-12-03)
+    * Generate CPU times of each CRC algorithm and variant using
+      [examples/benchmarks/CpuBenchmark](examples/benchmarks/CpuBenchmark).
+    * Create combined CPU and Memory benchmark tables in
+      [examples/benchmarks](examples/benchmarks).
+    * Insert a `crc_calculate()` convenience function into the header files so
+      that the CRC can be calculated in one-shot, instead of forcing the user to
+      call `crc_init()`, `crc_update()` and `crc_finalize()`.
+    * Remove `static` keyword from `*.hpp` files. Not needed in C++ and
+      prevents doxygen doc generation of those functions.
 * 0.1.1 (2020-11-29)
     * Fix embarrassing typos in library.properties.
 * 0.1 (2020-11-29)